Rework to use GitBook instead of MkDocs
The content was not reworked yet but just organized in a way GitBook
knows how to generate the book.

This commit changes the Dockerfile so it builds the HTML content and
exports it as usual. It allows for a seamless migration.

Now more polishment changes ought to be done in the content.

Change-Id: I841df2582364070f467549887196f84b2a3323ba
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>